gitsu_etl/requirements_py3.txt
pytz
PyYaml
ipython
github3.py
sqlalchemy
psycopg2-binary
